ref(uptime): Move ConnectedHttpSnippet into Detect panel
Fixes: [NEW-585: HTTP check details should be embedded in the detect panel](https://linear.app/getsentry/issue/NEW-585/http-check-details-should-be-embedded-in-the-detect-panel)
