+++ static/app/views/detectors/components/forms/uptime/detect/index.tsx
@@ -9,10 +9,12 @@ import SelectField from 'sentry/components/forms/fields/selectField';
 import TextareaField from 'sentry/components/forms/fields/textareaField';
 import TextField from 'sentry/components/forms/fields/textField';
 import type FormModel from 'sentry/components/forms/model';
+import {useFormField} from 'sentry/components/workflowEngine/form/useFormField';
 import {Container} from 'sentry/components/workflowEngine/ui/container';
 import Section from 'sentry/components/workflowEngine/ui/section';
 import {t, tct} from 'sentry/locale';
 import getDuration from 'sentry/utils/duration/getDuration';
+import {HTTPSnippet} from 'sentry/views/alerts/rules/uptime/httpSnippet';
 import {UptimeHeadersField} from 'sentry/views/detectors/components/forms/uptime/detect/uptimeHeadersField';
 import {UPTIME_DEFAULT_DOWNTIME_THRESHOLD} from 'sentry/views/detectors/components/forms/uptime/fields';
 
@@ -30,6 +32,30 @@ const VALID_INTERVALS_SEC = [
 
 function methodHasBody(model: FormModel) {
   return !HTTP_METHODS_NO_BODY.includes(model.getValue('method'));
+}
+
+function ConnectedHttpSnippet() {
+  const url = useFormField<string>('url');
+  const method = useFormField<string>('method');
+  const headers = useFormField<Array<[string, string]>>('headers');
+  const body = useFormField<string>('body');
+  const traceSampling = useFormField<boolean>('traceSampling');
+
+  if (!url || !method) {
+    return null;
+  }
+
+  const shouldIncludeBody = !HTTP_METHODS_NO_BODY.includes(method);
+
+  return (
+    <HTTPSnippet
+      url={url}
+      method={method}
+      headers={headers ?? []}
+      body={shouldIncludeBody ? (body ?? null) : null}
+      traceSampling={traceSampling ?? false}
+    />
+  );
 }
 
 export function UptimeDetectorFormDetectSection() {
@@ -152,6 +178,7 @@ export function UptimeDetectorFormDetectSection() {
             flexibleControlStateSize
           />
         </ConfigurationFieldsContainer>
+        <ConnectedHttpSnippet />
       </Section>
     </Container>
   );
